VisualStudio fixes

+++ src/main-conf.c
@@ -1602,7 +1602,7 @@ masscan_set_parameter(struct Masscan *masscan,
         exit(1);
     } else if (EQUALS("vlan", name) || EQUALS("adapter-vlan", name)) {
         masscan->nic[index].is_vlan = 1;
-        masscan->nic[index].vlan_id = parseInt(value);
+        masscan->nic[index].vlan_id = (unsigned)parseInt(value);
     } else if (EQUALS("wait", name)) {
         if (EQUALS("forever", value))
             masscan->wait =  INT_MAX;
